Fix: Security paths
