+++ Backend/controller/PersonController.go
@@ -38,7 +38,17 @@ func GetPersonById(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	utils.ReturnResponseJSON(w, http.StatusOK, "Pessoa encontrada com sucesso", person)
+	var personDto dtos.PersonDTO = dtos.PersonDTO{
+		IdPerson:  person.IdPerson,
+		Name:      person.Name,
+		BornDate:  person.BornDate,
+		DocNumber: person.DocNumber,
+		DocType:   person.DocType,
+		Password:  "",
+		Salt:      "",
+	}
+
+	utils.ReturnResponseJSON(w, http.StatusOK, "Pessoa encontrada com sucesso", personDto)
 }
 
 func PostPerson(w http.ResponseWriter, r *http.Request) {
@@ -85,7 +95,21 @@ func GetAllPerson(w http.ResponseWriter, _ *http.Request) {
 		return
 	}
 
-	utils.ReturnResponseJSON(w, http.StatusOK, "Busca realizada com sucesso!", persons)
+	var personsDto []dtos.PersonDTO
+
+	for i := 0; i < len(persons); i++ {
+		personsDto = append(personsDto, dtos.PersonDTO{
+			IdPerson:  persons[i].IdPerson,
+			Name:      persons[i].Name,
+			BornDate:  persons[i].BornDate,
+			DocNumber: persons[i].DocNumber,
+			DocType:   persons[i].DocType,
+			Password:  "",
+			Salt:      "",
+		})
+	}
+
+	utils.ReturnResponseJSON(w, http.StatusOK, "Busca realizada com sucesso!", personsDto)
 }
 
 func PutPerson(w http.ResponseWriter, r *http.Request) {
